Move generally-useful classes into spec helper.
