+++ Rakefile
@@ -72,9 +72,9 @@ task :console do
 end
 
 task :test do
-  require 'texticle_spec'
+  # require 'texticle_spec'
   require 'texticle/searchable_spec'
-  require 'texticle/full_text_indexer_spec'
+  # require 'texticle/full_text_indexer_spec'
 end
 
 namespace :db do
