+++ spec/texticle/full_text_indexer_spec.rb
@@ -1,6 +1,5 @@
 require 'spec_helper'
 require 'fileutils'
-require 'texticle/searchable'
 
 class FullTextIndexerTest < Test::Unit::TestCase
   context ".stream_output" do
